fix: :bug: fix all open issues with rpo-calculator (now proximity-ops)
