+++ src/locales/keys.ts
@@ -245,10 +245,23 @@ export const Keys = [
   'plugins.GraphicsMenuPlugin.bottomIconLabel',
   'plugins.GraphicsMenuPlugin.title',
   'plugins.GraphicsMenuPlugin.helpBody',
-  'plugins.RPOCalculator.bottomIconLabel',
-  'plugins.RPOCalculator.title',
-  'plugins.RPOCalculator.titleSecondary',
-  'plugins.RPOCalculator.helpBody',
+  'plugins.ProximityOps.bottomIconLabel',
+  'plugins.ProximityOps.title',
+  'plugins.ProximityOps.titleSecondary',
+  'plugins.ProximityOps.helpBody',
+  'plugins.ProximityOps.noradId',
+  'plugins.ProximityOps.maxDistThreshold',
+  'plugins.ProximityOps.maxRelativeVelocity',
+  'plugins.ProximityOps.searchDuration',
+  'plugins.ProximityOps.geoText',
+  'plugins.ProximityOps.leoText',
+  'plugins.ProximityOps.orbitType',
+  'plugins.ProximityOps.geoAllVsAll',
+  'plugins.ProximityOps.geoAllVsAllTooltip',
+  'plugins.ProximityOps.comparePayloadsOnly',
+  'plugins.ProximityOps.comparePayloadsOnlyTooltip',
+  'plugins.ProximityOps.ignoreVimpelRso',
+  'plugins.ProximityOps.ignoreVimpelRsoTooltip',
 ] as const;
 
 // Type for all valid translation keys
