+++ src/locales/locales.ts
@@ -95,10 +95,23 @@ export const loadLocalization = () => ({
       title: t7e('plugins.SensorTimeline.title'),
       helpBody: t7e('plugins.SensorTimeline.helpBody'),
     },
-    RPOCalculator: {
-      bottomIconLabel: i18next.t('plugins.RPOCalculator.bottomIconLabel'),
-      title: i18next.t('plugins.RPOCalculator.title'),
-      helpBody: i18next.t('plugins.RPOCalculator.helpBody'),
+    ProximityOps: {
+      bottomIconLabel: i18next.t('plugins.ProximityOps.bottomIconLabel'),
+      title: i18next.t('plugins.ProximityOps.title'),
+      helpBody: i18next.t('plugins.ProximityOps.helpBody'),
+      noradId: i18next.t('plugins.ProximityOps.noradId'),
+      maxDistThreshold: i18next.t('plugins.ProximityOps.maxDistThreshold'),
+      maxRelativeVelocity: i18next.t('plugins.ProximityOps.maxRelativeVelocity'),
+      searchDuration: i18next.t('plugins.ProximityOps.searchDuration'),
+      geoText: i18next.t('plugins.ProximityOps.geoText'),
+      leoText: i18next.t('plugins.ProximityOps.leoText'),
+      orbitType: i18next.t('plugins.ProximityOps.orbitType'),
+      geoAllVsAll: i18next.t('plugins.ProximityOps.geoAllVsAll'),
+      geoAllVsAllTooltip: i18next.t('plugins.ProximityOps.geoAllVsAllTooltip'),
+      comparePayloadsOnly: i18next.t('plugins.ProximityOps.comparePayloadsOnly'),
+      comparePayloadsOnlyTooltip: i18next.t('plugins.ProximityOps.comparePayloadsOnlyTooltip'),
+      ignoreVimpelRso: i18next.t('plugins.ProximityOps.ignoreVimpelRso'),
+      ignoreVimpelRsoTooltip: i18next.t('plugins.ProximityOps.ignoreVimpelRsoTooltip'),
     },
     SatelliteTimeline: {
       bottomIconLabel: t7e('plugins.SatelliteTimeline.bottomIconLabel'),
