+++ src/plugins/plugins.ts
@@ -44,8 +44,8 @@ import { Lat2LonPlots } from './plot-analysis/lat2lon';
 import { RicPlot } from './plot-analysis/ric-plots';
 import { Time2LonPlots } from './plot-analysis/time2lon';
 import { PolarPlotPlugin } from './polar-plot/polar-plot';
+import { ProximityOps } from './proximity-ops/proximity-ops';
 import { ReportsPlugin } from './reports/reports';
-import { RPOCalculator } from './rpo-calculator/rpo-calculator';
 import { SatConstellations } from './sat-constellations/sat-constellations';
 import { SatelliteFov } from './satellite-fov/satellite-fov';
 import { SatellitePhotos } from './satellite-photos/satellite-photos';
@@ -162,7 +162,7 @@ export const loadPlugins = (keepTrackApi: KeepTrackApi, plugins: KeepTrackPlugin
       { init: () => new PolarPlotPlugin().init(), enabled: plugins.polarPlot },
       { init: () => new NextLaunchesPlugin().init(), enabled: plugins.nextLaunch },
       { init: () => new FindSatPlugin().init(), enabled: plugins.findSat },
-      { init: () => new RPOCalculator().init(), enabled: plugins.RPOCalculator },
+      { init: () => new ProximityOps().init(), enabled: plugins.RPOCalculator },
       { init: () => new ShortTermFences().init(), enabled: plugins.shortTermFences },
       { init: () => new OrbitReferences().init(), enabled: plugins.orbitReferences },
       { init: () => new Collisions().init(), enabled: plugins.collisions },
